# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreateUserRequest:
    properties:
      email:
        description: 邮箱（推荐填写，用于两步验证重置）
        example: lisi@example.com
        type: string
      nickname:
        description: 昵称
        example: 李四
        type: string
      password:
        description: 密码（至少6位）
        example: secret123
        type: string
      role:
        description: 角色：user | admin（默认 user）
        enum:
        - user
        - admin
        example: user
        type: string
      username:
        description: 用户名
        example: lisi
        type: string
    required:
    - password
    - username
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: 'failed to read uploaded image: ...'
        type: string
    type: object
...
  handler.UpdateUserRequest:
    properties:
      email:
        description: 邮箱（可选，本人或管理员可改）
        example: zhangsan@example.com
        type: string
      nickname:
        description: 昵称（可选）
        example: 张三
        type: string
      role:
        description: 角色（仅管理员可改）
        enum:
        - user
        - admin
        example: user
        type: string
    type: object
  handler.UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/user.SafeUser'
        type: array
      total:
        type: integer
    type: object
...
  user.SafeUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      mfa_enabled:
        description: 是否已开启两步验证
        type: boolean
      nickname:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: wechatapp-web
    url: http://www.swagger.io/support
  description: |-
    Image background replacement API. Uploads an image, removes its
    background via the remove.bg service, then optionally composites
    the cutout onto a new background (solid color or another image).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: wechatapp-web API
  version: 1.0.0
paths:
...
  /users:
    get:
      description: Returns all accounts. Admin only.
      produces:
      - application/json
      responses:
        "200":
          description: User list
          schema:
            $ref: '#/definitions/handler.UserListResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Not an admin
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users (admin)
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates an account with an explicit role. Admin only.
      parameters:
      - description: New account
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created user
          schema:
            $ref: '#/definitions/user.SafeUser'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Not an admin
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Username already exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a user (admin)
      tags:
      - users
  /users/{id}:
    delete:
      description: Removes an account. Admin only.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: 用户已删除'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Not an admin
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user (admin)
      tags:
      - users
    get:
      description: Returns one account. Admin, or the user themselves.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User detail
          schema:
            $ref: '#/definitions/user.SafeUser'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Not allowed to view this user
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user detail
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates nickname (self or admin) and role (admin only).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user
          schema:
            $ref: '#/definitions/user.SafeUser'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Not allowed to edit this user
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT issued by POST /auth/login, sent as "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"